experimental: do not depend on Oscar
This is a version which can be used either with Oscar.jl or with Gapjm.jl.
The idea is to use Oscar.jl's polynomials if this package is already loaded
before GenericDecMats.jl, and to use Gapjm.jl's multivariate polynomials
if this package is already loaded.

(For technical reasons, Requires.jl is used.
However, we cannot use its feature that the functionality depending on
one of the two packages becomes available as soon as Oscar.jl or Gapjm.jl
gets loaded later on.
Note that the type `GenericDecompositionMatrix` can support only one
of the two types of polynomials.)